Add license information to many files. PCB design files are CC-BY and code is MIT license.

+++ test/Src/main.cpp
@@ -30,6 +30,29 @@
   *
   ******************************************************************************
   */
+
+// MIT License
+//
+// Copyright (c) 2016 Chickadee Tech LLC
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 /* Includes ------------------------------------------------------------------*/
 #include <inttypes.h>
 #include <tuple>
@@ -629,7 +652,7 @@ void i2cWrite(char* line) {
   printf("ok\n");
   HAL_Delay(1);
 
-  uint8_t offset = 0;
+  uint16_t offset = 0;
   while (offset < totalBytesToWrite) {
     char line[128];
     char* line_status = fgets(line, 128, stdin);
